+++ src/internal/internal.h
@@ -9,6 +9,11 @@ extern sexp* rlang_zap;
 extern sexp* as_list_call;
 extern sexp* as_list_s4_call;
 
+extern sexp* rlang_objs_keep;
+extern sexp* rlang_objs_trailing;
+
+extern sexp* fns_function;
+
 void rlang_init_internal();
 sexp* rlang_ns_get(const char* name);